pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "darwin"
authors = [{name = "Archer Heffern", email = "heffernarcher@gmail.com"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = ["License :: OSI Approved :: MIT License"]
dynamic = ["version", "description"]
dependencies = [
"requests==2.32.3",
"python-dotenv==1.0.1",
"fastapi[standard]==0.115.0",
]

[project.optional-dependencies]
dev = [
    "black"
]

[project.urls]
Home = "https://github.com/archerheffern/DarwinAPI"